requirements.txt
aiogram==3.1.1
APScheduler==3.10.4
python-dotenv~=1.0.0